Move tests_path from Application class to Test class
For execution, each test now tracks the directory in which it is held.
This is not specified in the config card. This is determined when the
Test instances are constructed from their config files (we assume the
input file paths specified in these are relative to the config card
location).